CLT-23: Created LB configuration bash script

+++ CL_Project-Maks/Web-server_Configuration/ws_conf.sh
@@ -19,12 +19,15 @@ apt install nginx -y
 
 systemctl start nginx
 
-mv ws_conf_template /etc/nginx/sites-available
+cp ws_conf_template /etc/nginx/sites-available
 
 ln -s /etc/nginx/sites-available/ws_conf_template /etc/nginx/sites-enabled/
-rm /etc/nginx/sites-enabled/default
 
-cat ip_conf_template > /etc/netplan/50-cloud-init.yaml
+if [ -f "/etc/nginx/sites-enabled/default" ]; then
+    rm /etc/nginx/sites-enabled/default
+fi
+
+envsubst < ip_conf_template > /etc/netplan/50-cloud-init.yaml
 netplan apply
 
 systemctl restart nginx
@@ -43,6 +46,6 @@ sudo apt install python3 python3-pip python3-venv -y
 python3 -m venv ../bird_app/.venv
 . ../bird_app/.venv/bin/activate
 
-pip install -r requirements.txt
+pip install -r ../bird_app/requirements.txt
 
 python3 ../bird_app/app.py
